Organize files in filter directory
