+++ src/components/molecules/Filter/ItemData.js
@@ -1,0 +1,50 @@
+const CATEGORY = [
+  { key: 0, title: "웹", type: "category", link: "" },
+  { key: 1, title: "어플리케이션", type: "category", link: "" },
+  { key: 2, title: "게임", type: "category", link: "" },
+  { key: 3, title: "그래픽", type: "category", link: "" },
+  { key: 4, title: "브랜딩", type: "category", link: "" },
+  { key: 5, title: "소프트웨어", type: "category", link: "" },
+  { key: 6, title: "영상", type: "category", link: "" },
+  { key: 7, title: "임베디드", type: "category", link: "" },
+  { key: 8, title: "제품", type: "category", link: "" },
+  { key: 9, title: "출판", type: "category", link: "" },
+  { key: 10, title: "퍼블리싱", type: "category", link: "" },
+  { key: 11, title: "기타", type: "category", link: "" },
+];
+const FIELD = [
+  { key: 0, title: "기획", type: "field", link: "" },
+  { key: 1, title: "개발", type: "field", link: "" },
+  { key: 2, title: "디자인", type: "field", link: "" },
+];
+const REGION = [
+  { key: 0, title: "서울", type: "region", link: "" },
+  { key: 1, title: "경기", type: "region", link: "" },
+  { key: 2, title: "광주", type: "region", link: "" },
+  { key: 3, title: "대구", type: "region", link: "" },
+  { key: 4, title: "대전", type: "region", link: "" },
+  { key: 5, title: "부산", type: "region", link: "" },
+  { key: 6, title: "인천", type: "region", link: "" },
+  { key: 7, title: "울산", type: "region", link: "" },
+  { key: 8, title: "세종", type: "region", link: "" },
+  { key: 9, title: "제주", type: "region", link: "" },
+  { key: 10, title: "강원", type: "region", link: "" },
+  { key: 11, title: "경남", type: "region", link: "" },
+  { key: 12, title: "경북", type: "region", link: "" },
+  { key: 13, title: "전남", type: "region", link: "" },
+  { key: 14, title: "전북", type: "region", link: "" },
+  { key: 15, title: "충남", type: "region", link: "" },
+  { key: 16, title: "충북", type: "region", link: "" },
+];
+const PROJECTTYPE = [
+  { key: 0, title: "공모전", type: "projecttype", link: "" },
+  { key: 1, title: "창업", type: "projecttype", link: "" },
+  { key: 2, title: "프로젝트", type: "projecttype", link: "" },
+  { key: 3, title: "기타", type: "projecttype", link: "" },
+];
+
+const ALIGN = [
+  { id: 0, title: "조회순", key: "align", link: "" },
+  { id: 1, title: "댓글순", key: "align", link: "" },
+];
+export { CATEGORY, FIELD, REGION, ALIGN, PROJECTTYPE };
