+++ src/components/organisms/FilterSearch.js
@@ -2,47 +2,59 @@ import styled from "styled-components";
 import Search from "../molecules/Input/Search";
 import Filter from "../molecules/Filter/Filter";
 import Wrapper from "../atoms/FilterSearch";
-import {PROJECTTYPE, FIELD, REGION, CATEGORY} from "../molecules/Filter/Item"
+import {
+  PROJECTTYPE,
+  FIELD,
+  REGION,
+  CATEGORY,
+} from "../molecules/Filter/ItemData";
 export default function FilterSearch(props) {
-	
-	const category = "카테고리";
-	const region = "지역";
-	const field = "구인분야";
-	const projecttype = "프로젝트 종류";
+  const category = "카테고리";
+  const region = "지역";
+  const field = "구인분야";
+  const projecttype = "프로젝트 종류";
 
-	return (
-		<Wrapper>
-			<InnerWrapper>
-				<FilterWrapper>
-					<Filter title={category} activeMenu="category" data={CATEGORY}></Filter>
-					<Filter title={field} activeMenu="field" data={FIELD}></Filter>
-					{props.type === "project" && (
-						<>
-							<Filter title={region} activeMenu="region" data={REGION}></Filter>
-							<Filter title={projecttype} activeMenu="projecttype" data={PROJECTTYPE}></Filter>
-						</>
-					)}
-				</FilterWrapper>
-				<Search></Search>
-			</InnerWrapper>
-		</Wrapper>
-	);
+  return (
+    <Wrapper>
+      <InnerWrapper>
+        <FilterWrapper>
+          <Filter
+            title={category}
+            activeMenu="category"
+            data={CATEGORY}
+          ></Filter>
+          <Filter title={field} activeMenu="field" data={FIELD}></Filter>
+          {props.type === "project" && (
+            <>
+              <Filter title={region} activeMenu="region" data={REGION}></Filter>
+              <Filter
+                title={projecttype}
+                activeMenu="projecttype"
+                data={PROJECTTYPE}
+              ></Filter>
+            </>
+          )}
+        </FilterWrapper>
+        <Search></Search>
+      </InnerWrapper>
+    </Wrapper>
+  );
 }
 
 const FilterWrapper = styled.div`
-	width: 23rem;
-	height: 1.6rem;
-	justify-content: left;
-	align-items: center;
-	display: flex;
-	flex-direction: row;
-	margin: 0 15rem 0 0rem;
+  width: 23rem;
+  height: 1.6rem;
+  justify-content: left;
+  align-items: center;
+  display: flex;
+  flex-direction: row;
+  margin: 0 15rem 0 0rem;
 `;
 const InnerWrapper = styled.div`
-	display: flex;
-	flex-direction: row;
-	box-sizing: border-box;
-	margin: 0 4rem 0 4rem;
-	max-width: 1200px;
-	width: 48rem;
+  display: flex;
+  flex-direction: row;
+  box-sizing: border-box;
+  margin: 0 4rem 0 4rem;
+  max-width: 1200px;
+  width: 48rem;
 `;
