+++ src/components/templates/ProjectBody.js
@@ -2,48 +2,48 @@ import ProjectList from "./ProjectList";
 import Filter from "../molecules/Filter/Filter";
 import WriteButton from "../molecules/Button/Write";
 import styled from "styled-components";
-import {ALIGN} from "../molecules/Filter/Item"
+import { ALIGN } from "../molecules/Filter/ItemData";
+
 export default function ProjectFilter() {
-	
-	return (
-		<Wrapper>
-			<InnerWrapper>
-				<Div>
-					<Filter title="최신순" activeMenu="align" data={ALIGN}></Filter>
-				</Div>
-				<ProjectList></ProjectList>
-			</InnerWrapper>
-			<WriteButton></WriteButton>
-		</Wrapper>
-	);
+  return (
+    <Wrapper>
+      <InnerWrapper>
+        <Div>
+          <Filter title="최신순" activeMenu="align" data={ALIGN}></Filter>
+        </Div>
+        <ProjectList></ProjectList>
+      </InnerWrapper>
+      <WriteButton></WriteButton>
+    </Wrapper>
+  );
 }
 
 const Div = styled.div`
-	width: 100%;
-	display: flex;
-	flex-direction: row;
-	justify-content: flex-end;
-	margin: 1.2rem 0 1.2rem 1rem;
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-end;
+  margin: 1.2rem 0 1.2rem 1rem;
 `;
 
 const InnerWrapper = styled.div`
-	margin: 0 4rem 0 4rem;
-	max-width: 1200px;
-	width: 48rem;
-	flex-direction: column;
-	justify-content: space-between;
-	align-items: center;
-	box-sizing: border-box;
-	position: relative;
-	z-index: 100;
+  margin: 0 4rem 0 4rem;
+  max-width: 1200px;
+  width: 48rem;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  box-sizing: border-box;
+  position: relative;
+  z-index: 100;
 `;
 
 const Wrapper = styled.div`
-	background-color: #f0f8fd;
-	justify-content: center;
-	width: 100%;
-	align-items: center;
-	display: flex;
-	flex-direction: row;
-	box-sizing: border-box;
+  background-color: #f0f8fd;
+  justify-content: center;
+  width: 100%;
+  align-items: center;
+  display: flex;
+  flex-direction: row;
+  box-sizing: border-box;
 `;
